switch to HTML templates
