--- views/template.html
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-  <head>
-
-    <title>Late Night Menu</title>
-
-    <style>
-      body {
-        font-family: 'Helvetica Neue', Helvetica, Arial, sans-serif;
-      }
-    </style>
-
-  </head>
-
-  <body>
-
-    <h1>Late Night Menu</h1>
-      [MENU_PLACEHOLDER]
-
-  </body>
-
-</html>
